requirements.txt
absl-py==1.3.0
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
datasets==2.8.0
dill==0.3.6
Django==4.1.4
evaluate==0.4.0
filelock==3.8.2
Flask==2.3.3
frozenlist==1.3.3
fsspec==2022.11.0
huggingface-hub==0.11.1
idna==3.4
inflate64==0.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
nltk==3.8
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
pandas==1.5.2
protobuf==4.21.12
psutil==5.9.4
py7zr==0.20.2
pyarrow==10.0.1
pybcj==1.0.1
pycryptodomex==3.16.0
pyppmd==1.0.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.7
PyYAML==6.0
pyzstd==0.15.3
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rouge-score==0.1.2
sentencepiece==0.1.97
sexpdata==1.0.1
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.3.7
xxhash==3.1.0
yarl==1.8.2
